Move subscription primitives to package root
models/ is the data layer; SubscriptionHandler is a stateful
infrastructure mixin used by both legacy and v1 handler trees.
Move it to aiounifi/subscription.py (alongside controller.py and
errors.py) where cross-cutting infrastructure belongs.

Also:
- Drop unused ABC base; SubscriptionHandler has no abstract methods
- Fix unbound _id_filter variable for tuple id_filter inputs in subscribe()
- Tighten type annotation: tuple[str] -> tuple[str, ...]
